Introduce proper support for unread files.
Largely replaces --retour functionality. Unread (-u) marks files as
unread upon first land. Those files do not show up in rover tour (unlike
--retour). Unread files can be fetched using 'rover fetch . -u' while in
a rover directory. This is such a common operation that 'rover fetch' is
a shortcut for 'rover fetch . -u'.

I may remove the 'rover fetch' shortcut in the future as it should be up
to the user to choose aliases.

Other minor changes in this revision:
* Any errors upon land are now propogated out of rover using 'error_occurred'.
* Fix a problem where mark_for_land would err when attempting to delete
  'local_path' when 'local_path' did not exist. This is a problem when
  using the 'rover land -u' feature.